rename and pod structure mlc report components
