+++ app/benchmarking/components/reportComponents/benchmarkReportRecipeResult.tsx
@@ -6,7 +6,10 @@ import {
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { colors } from '@/app/customColors';
 import { RecipeGradeBadge } from './badge';
-import { gradingLettersMlcMap, MLC_COOKBOOK_IDS } from './constants';
+import {
+  gradingLettersMlcMap,
+  MLC_COOKBOOK_IDS,
+} from './mlcReportComponents/constants';
 import { gradeColorsMlc, gradeColorsMoonshot } from './gradeColors';
 import { RangedBarChart } from './rangedBarChart';
