+++ app/benchmarking/components/reportComponents/benchmarkReportSectionTwo.tsx
@@ -5,7 +5,7 @@ import { LoadingAnimation } from '@/app/components/loadingAnimation';
 import { useGetCookbooksQuery } from '@/app/services/cookbook-api-service';
 import { SquareBadge } from './badge';
 import { BenchmarkReportCookbookResult } from './benchmarkReportCookbookResult';
-import { MLC_COOKBOOK_IDS } from './constants';
+import { MLC_COOKBOOK_IDS } from './mlcReportComponents/constants';
 import { gradeColorsMoonshot } from './gradeColors';
 
 type BenchmarkReportProps = {
