+++ app/benchmarking/components/reportComponents/mlcReportComponents/gradeColors.ts
@@ -1,0 +1,9 @@
+import { GradingColors } from '@/app/benchmarking/types/benchmarkReportTypes';
+
+export const gradeColorsMlc: GradingColors = {
+  'Low Risk': '#25A167',
+  'Moderate-Low Risk': '#669C44',
+  'Moderate Risk': '#7B94BD',
+  'Moderate-High Risk': '#A19227',
+  'High Risk': '#C4805E',
+};
